// src/event-logger/types.ts
export enum EventType {
  CREATE_GAME = 'CREATE_GAME',
  JOIN_GAME = 'JOIN_GAME',
  ADD_AI = 'ADD_AI',
  LEAVE_GAME = 'LEAVE_GAME',
  START_GAME = 'START_GAME',
  PLACE_ARMIES = 'PLACE_ARMIES',
  END_TURN = 'END_TURN',
  MOVE_ARMIES = 'MOVE_ARMIES',
  ATTACK = 'ATTACK',
  WIN_CONTINENT = 'WIN_CONTINENT',
  ADD_CARD = 'ADD_CARD',
  USE_CARD = 'USE_CARD',
  SURRENDER = 'SURRENDER',
  ELIMINATE_PLAYER = 'ELIMINATE_PLAYER',
  GENERATE_ARMIES = 'GENERATE_ARMIES',

  BOT_GET_ENEMY_ZONES = 'BOT_GET_ENEMY_ZONES',
  BOT_GET_MY_ZONES = 'BOT_GET_MY_ZONES',
  BOT_ZONE_ENEMIES = 'BOT_ZONE_ENEMIES',
  BOT_ZONES_WITH_ENEMIES = 'BOT_ZONES_WITH_ENEMIES',
  BOT_ATTACK = 'BOT_ATTACK',
  BOT_PLACE_ARMIES = 'BOT_PLACE_ARMIES',
}